git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16548 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/VariableClassificationBuilder.java
@@ -325,7 +325,7 @@ public class VariableClassificationBuilder {
   private void dumpVariableTypeMapping(Path target, VariableClassification vc) {
     try (Writer w = Files.openOutputFile(target)) {
         for (String var : allVars) {
-          byte type = 0;
+          int type = 0;
           if (vc.getIntBoolVars().contains(var)) {
             type += 1 + 2 + 4; // IntBool is subset of IntEqualBool and IntAddEqBool
           } else if (vc.getIntEqualVars().contains(var)) {
